+++ neural_trees/classical/naive_bayes.py
@@ -14,8 +14,12 @@ from sklearn.base import BaseEstimator, ClassifierMixin
 from sklearn.preprocessing import LabelEncoder
 from sklearn.utils.validation import check_X_y, check_array, check_is_fitted
 
+from sklearn.utils.multiclass import check_classification_targets
 
-class NaiveBayesClassifier(BaseEstimator, ClassifierMixin):
+from neural_trees._validation import check_predict_input
+
+
+class NaiveBayesClassifier(ClassifierMixin, BaseEstimator):
     """
     Naive Bayes Classifier with selectable likelihood.
 
@@ -40,6 +44,7 @@ class NaiveBayesClassifier(BaseEstimator, ClassifierMixin):
 
     def fit(self, X, y):
         X, y = check_X_y(X, y)
+        check_classification_targets(y)
         self.le_ = LabelEncoder()
         y_enc = self.le_.fit_transform(y)
         self.classes_ = self.le_.classes_
@@ -86,7 +91,7 @@ class NaiveBayesClassifier(BaseEstimator, ClassifierMixin):
 
     def predict_log_proba(self, X):
         check_is_fitted(self)
-        X = check_array(X)
+        X = check_predict_input(self, X)
         log_probs = np.column_stack([
             self.class_log_prior_[c] + self._log_likelihood(X, c)
             for c in range(len(self.classes_))
@@ -100,5 +105,6 @@ class NaiveBayesClassifier(BaseEstimator, ClassifierMixin):
         return probs / probs.sum(axis=1, keepdims=True)
 
     def predict(self, X):
+        check_is_fitted(self)
         check_is_fitted(self)
         return self.le_.inverse_transform(self.predict_log_proba(X).argmax(axis=1))